Fix spinner height to maximum at 45deg
Because the spinner element rotates it would behave poorly when
displayed in a context where it contributes to the height of a parent
element, especially if that parent element scrolls. By fixing the height
at the maximum height the spinner will become when rotated to e.g. 45deg
we ensure that it's always the same size.